Changing the argparse custom argument types.
Previously we had a file_path and file_path_endswith_md custom
types. These are now combined into a file_path_endswith custom type
which is parameterized with the endswith argument. The same custom
type checks if the given parameter is a file, no specific extension.
Or if a specific extension is also given, check if the parameter is a
file with the user specified extension.