Bug with pandas to_markdown when the csv cells contain newline.
When the cells contain newlines, multi-paragraph text, the pandas
to_markdown does not create the markdown table as expected. When the
newlines are replaced with the html <br> tag the output is as
expected. Added a utilities function which does the replacement as
part of the conversion to markdown and replaced the original
to_markdown calls with this utility function.