Adding script for validating the bibliography file.
Added script for validating bibliography style, including enforcing
some knowledge-base specific requirements (doi and notes keys). Also
includes accompanying testing. This also lead to an update to the
bibliography test file which didn't include the notes section. As a
result the md5 hash for the bibliography to markdown script had to be
updated too.