Removing superfluous argparse related code.
No need for manually removing first argument from sys.argv and passing
that list to parse_args. Just passing the given function parameter which has a
default of None is sufficient. When argparse receives args=None as
input it will use the results of calling sys.argv, otherwise it uses
the args list (useful for debugging).