Adding empty cell and image usage checks.
Check that there are no empty cells in the reagent_resources.csv file.
If there is no information for a cell it should contain the string
"NA". Contributors need to explicitly specify a value,
reducing the likelihood of an erroneously empty cell.

Check that images and captions listed in the reagent_resources.csv
file are referenced in the corresponding supporting material files.