Update readme with guidance on how to run the scripts.
Replaced the previous, readme specific, instructions with hyperlinks
to the KB workflow usage of the utility scripts. This ensures that there is a
single authoritative example, the automated workflow scripts which are always
up to date. Otherwise, every change to the utilities scripts would require
updating the workflow scripts and readme.