Improve URL checking.
Improving URL checking by using sessions for same domain and
attempting to mimic a browser so increase chances of not getting 403
error. This is not a comprehensive solution and still doesn't work for
many sites that have DOS protection in place and do not want
scripts/bots to access them.